Reorganize verb subcomponents
