+++ javascript/react/src/compounds/CompoundBinarize/CompoundBinarize.tsx
@@ -7,9 +7,10 @@ import { ActionButton, TextField } from '@fluentui/react'
 import { memo, useMemo, useCallback, useState } from 'react'
 import styled from 'styled-components'
 import { LeftAlignedRow, useLoadTable } from '../../common/index.js'
-import { FilterInputs, TableColumnDropdown } from '../../controls/index.js'
+import { TableColumnDropdown } from '../../controls/index.js'
 import { dropdownStyles } from '../../controls/styles.js'
 import type { StepComponentProps } from '../../types.js'
+import { FilterInputs } from '../../verbs/shared/index.js'
 import {
 	createBinarize,
 	defaults,
