+++ javascript/react/src/compounds/FilterAggregateLookup/FilterAggregateLookup.tsx
@@ -8,13 +8,13 @@ import styled from 'styled-components'
 import { LeftAlignedRow, useLoadTable } from '../../common/index.js'
 import {
 	FieldAggregateOperationDropdown,
-	FilterFunction,
 	ReadOnlyTextField,
 	TableColumnDropdown,
 	TableDropdown,
 } from '../../controls/index.js'
 import { dropdownStyles } from '../../controls/styles.js'
 import type { StepComponentProps } from '../../types.js'
+import { FilterFunction } from '../../verbs/shared/index.js'
 import {
 	getAggregate,
 	defaults,
