--- javascript/react/src/controls/RollupInputs/index.ts
@@ -1,5 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from './RollupInputs.js'
