--- javascript/react/src/controls/TableListInputs/TableListInputs.tsx
@@ -1,100 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import type {
-	TableStore,
-	SetOperationStep,
-	Step,
-} from '@data-wrangling-components/core'
-import { ActionButton, IconButton } from '@fluentui/react'
-import { memo, useCallback, useMemo } from 'react'
-import styled from 'styled-components'
-import { LeftAlignedRow, useLoadTable } from '../../common/index.js'
-import type { StepComponentProps } from '../../types.js'
-import { TableDropdown } from '../index.js'
-
-/**
- * Provides inputs to create a list of tables.
- * E.g., for set operations
- */
-export const TableListInputs: React.FC<StepComponentProps> = memo(
-	function TableListInputs({ step, store, table, onChange, input }) {
-		const internal = useMemo(() => step as SetOperationStep, [step])
-
-		const tbl = useLoadTable(input || internal.input, table, store)
-
-		const others = useOthers(internal, store, onChange)
-
-		const handleButtonClick = useCallback(() => {
-			onChange &&
-				onChange({
-					...internal,
-					args: {
-						...internal.args,
-						// TODO: this just establishes an empty table, can we get the store list and pick the next available?
-						others: [...internal.args.others, ''],
-					},
-				})
-		}, [internal, onChange])
-
-		return (
-			<Container>
-				{others}
-				<ActionButton
-					onClick={handleButtonClick}
-					iconProps={{ iconName: 'Add' }}
-					disabled={!tbl}
-				>
-					Add table
-				</ActionButton>
-			</Container>
-		)
-	},
-)
-
-function useOthers(
-	step: SetOperationStep,
-	store?: TableStore,
-	onChange?: (step: Step) => void,
-) {
-	return useMemo(() => {
-		return step.args.others.map((other, index) => {
-			const handleDeleteClick = () => {
-				const update = { ...step }
-				update.args.others.splice(index, 1)
-				onChange && onChange(update)
-			}
-			if (!store) {
-				return null
-			}
-			return (
-				<LeftAlignedRow key={`set-op-${other}-${index}`}>
-					<TableDropdown
-						label={''}
-						store={store}
-						selectedKey={other}
-						onChange={(_evt, option) => {
-							const update = { ...step }
-							if (option) {
-								update.args.others[index] = `${option.key}`
-							}
-							onChange && onChange(update)
-						}}
-					/>
-					<IconButton
-						title={'Remove this table'}
-						iconProps={{ iconName: 'Delete' }}
-						onClick={handleDeleteClick}
-					/>
-				</LeftAlignedRow>
-			)
-		})
-	}, [step, store, onChange])
-}
-
-const Container = styled.div`
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-`
