+++ javascript/react/src/controls/index.ts
@@ -2,23 +2,9 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './FoldInputs/index.js'
-export * from './EraseInputs/index.js'
-export * from './MergeInputs/index.js'
-export * from './PivotInputs/index.js'
-export * from './UnfoldInputs/index.js'
-export * from './ImputeInputs/index.js'
-export * from './SpreadInputs/index.js'
+export * from './ColumnOrValueComboBox.js'
 export * from './ColumnInstruction.js'
-export * from './ColumnListInputs/index.js'
-export * from './JoinInputs/index.js'
-export * from './FilterInputs/index.js'
-export * from './AggregateInputs/index.js'
-export * from './ImputeInputs/index.js'
-export * from './RollupInputs/index.js'
-export * from './FilterFunction/index.js'
 export * from './ReadOnlyTextField.js'
 export * from './SortInstruction.js'
-export * from './TableListInputs/index.js'
 export * from './dropdowns/index.js'
 export * from './GroupHeader/index.js'
