+++ javascript/react/src/verbs/Aggregate/Aggregate.tsx
@@ -2,20 +2,68 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { memo } from 'react'
-import { VerbContainer } from '../../common/index.js'
-import { AggregateInputs } from '../../controls/index.js'
+import type { AggregateStep } from '@data-wrangling-components/core'
+import { memo, useMemo } from 'react'
+import styled from 'styled-components'
+import {
+	useLoadTable,
+	LeftAlignedRow,
+	useHandleDropdownChange,
+} from '../../common/index.js'
+import {
+	FieldAggregateOperationDropdown,
+	TableColumnDropdown,
+} from '../../controls/index.js'
 import type { StepComponentProps } from '../../types.js'
 
 /**
- * Provides inputs for an aggregation step.
+ * Just the group/column/op inputs for an aggregation.
+ * Input table is expected to be edited elsewhere and configured as the step input.
  */
-export const Aggregate: React.FC<StepComponentProps> = memo(function Aggregate(
-	props,
-) {
+export const Aggregate: React.FC<StepComponentProps> = memo(function Aggregate({
+	step,
+	store,
+	table,
+	onChange,
+	input,
+}) {
+	const internal = useMemo(() => step as AggregateStep, [step])
+
+	const tbl = useLoadTable(input || internal.input, table, store)
+
+	const handleGroupColumnChange = useHandleDropdownChange(
+		internal,
+		'args.groupby',
+		onChange,
+	)
+
+	const handleOpChange = useHandleDropdownChange(
+		internal,
+		'args.operation',
+		onChange,
+	)
+
 	return (
-		<VerbContainer>
-			<AggregateInputs {...props} />
-		</VerbContainer>
+		<Container>
+			<LeftAlignedRow>
+				<TableColumnDropdown
+					required
+					table={tbl}
+					label={'Column to group by'}
+					selectedKey={internal.args.groupby}
+					onChange={handleGroupColumnChange}
+				/>
+				<FieldAggregateOperationDropdown
+					selectedKey={internal.args.operation}
+					onChange={handleOpChange}
+				/>
+			</LeftAlignedRow>
+		</Container>
 	)
 })
+
+const Container = styled.div`
+	display: flex;
+	flex-direction: row;
+	justify-content: flex-start;
+`
