+++ javascript/react/src/verbs/ColumnListOperation/ColumnListOperation.tsx
@@ -3,8 +3,8 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import { memo } from 'react'
-import { ColumnListInputs } from '../../controls/index.js'
 import type { StepComponentProps } from '../../types.js'
+import { ColumnListInputs } from '../shared/index.js'
 
 /**
  * Provides inputs for a ColumnListOperation step.
