+++ javascript/react/src/verbs/Join/Join.tsx
@@ -3,8 +3,8 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import { memo } from 'react'
-import { JoinInputs } from '../../controls/index.js'
 import type { StepComponentProps } from '../../types.js'
+import { JoinInputs } from '../shared/index.js'
 
 /**
  * Provides inputs for a Join step.
