+++ javascript/react/src/verbs/Pivot/Pivot.tsx
@@ -2,18 +2,85 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { memo } from 'react'
-import { VerbContainer } from '../../common'
-import { PivotInputs } from '../../controls'
+import type { PivotStep } from '@data-wrangling-components/core'
+import { memo, useMemo } from 'react'
+import styled from 'styled-components'
+import {
+	useLoadTable,
+	LeftAlignedRow,
+	useHandleDropdownChange,
+} from '../../common'
+import {
+	FieldAggregateOperationDropdown,
+	TableColumnDropdown,
+} from '../../controls/index.js'
 import type { StepComponentProps } from '../../types'
 
 /**
- * Provides inputs for an aggregation step.
+ * Just the group/column/op inputs for an aggregation.
+ * Input table is expected to be edited elsewhere and configured as the step input.
  */
-export const Pivot: React.FC<StepComponentProps> = memo(function Pivot(props) {
+export const Pivot: React.FC<StepComponentProps> = memo(function Pivot({
+	step,
+	store,
+	table,
+	onChange,
+	input,
+}) {
+	const internal = useMemo(() => step as PivotStep, [step])
+
+	const tbl = useLoadTable(input || internal.input, table, store)
+
+	const handleKeyColumnChange = useHandleDropdownChange(
+		internal,
+		'args.key',
+		onChange,
+	)
+
+	const handleValueColumnChange = useHandleDropdownChange(
+		internal,
+		'args.value',
+		onChange,
+	)
+
+	const handleOpChange = useHandleDropdownChange(
+		internal,
+		'args.operation',
+		onChange,
+	)
+
 	return (
-		<VerbContainer>
-			<PivotInputs {...props} />
-		</VerbContainer>
+		<Container>
+			<LeftAlignedRow>
+				<TableColumnDropdown
+					required
+					table={tbl}
+					label={'Column used as key'}
+					selectedKey={internal.args.key}
+					onChange={handleKeyColumnChange}
+				/>
+			</LeftAlignedRow>
+			<LeftAlignedRow>
+				<TableColumnDropdown
+					required
+					table={tbl}
+					label={'Column used as value'}
+					selectedKey={internal.args.value}
+					onChange={handleValueColumnChange}
+				/>
+				<FieldAggregateOperationDropdown
+					selectedKey={internal.args.operation}
+					onChange={handleOpChange}
+				/>
+			</LeftAlignedRow>
+		</Container>
 	)
 })
+
+const Container = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	align-content: flex-start;
+	justify-content: flex-start;
+`
