+++ javascript/react/src/verbs/Rollup/Rollup.tsx
@@ -2,20 +2,42 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { memo } from 'react'
-import { VerbContainer } from '../../common/index.js'
-import { RollupInputs } from '../../controls/index.js'
+import type { RollupStep } from '@data-wrangling-components/core'
+import { memo, useMemo } from 'react'
+import styled from 'styled-components'
+import { LeftAlignedRow, useHandleDropdownChange } from '../../common/index.js'
+import { FieldAggregateOperationDropdown } from '../../controls/index.js'
 import type { StepComponentProps } from '../../types.js'
-
 /**
- * Provides inputs for a rollup step.
+ * Just the column/op inputs for an rollup.
+ * Input table is expected to be edited elsewhere and configured as the step input.
  */
-export const Rollup: React.FC<StepComponentProps> = memo(function Rollup(
-	props,
-) {
+export const Rollup: React.FC<StepComponentProps> = memo(function Rollup({
+	step,
+	onChange,
+}) {
+	const internal = useMemo(() => step as RollupStep, [step])
+
+	const handleOpChange = useHandleDropdownChange(
+		internal,
+		'args.operation',
+		onChange,
+	)
+
 	return (
-		<VerbContainer>
-			<RollupInputs {...props} />
-		</VerbContainer>
+		<Container>
+			<LeftAlignedRow>
+				<FieldAggregateOperationDropdown
+					selectedKey={internal.args.operation}
+					onChange={handleOpChange}
+				/>
+			</LeftAlignedRow>
+		</Container>
 	)
 })
+
+const Container = styled.div`
+	display: flex;
+	flex-direction: row;
+	justify-content: flex-start;
+`
