+++ javascript/react/src/verbs/shared/ColumnListInputs/ColumnListInputs.tsx
@@ -2,30 +2,26 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type {
-	SpreadArgs,
-	SpreadStep,
-	Step,
-} from '@data-wrangling-components/core'
-import { ActionButton, Label } from '@fluentui/react'
+import type { ColumnListStep, Step } from '@data-wrangling-components/core'
+import { ActionButton } from '@fluentui/react'
 import type ColumnTable from 'arquero/dist/types/table/column-table'
 import set from 'lodash-es/set.js'
 import { memo, useCallback, useMemo } from 'react'
 import styled from 'styled-components'
-import { useHandleDropdownChange, useLoadTable } from '../../common'
-import { TableColumnDropdown } from '../../controls'
-import type { StepComponentProps } from '../../types'
-import { ColumnSpread } from '../ColumnSpread'
+import { ColumnInstruction } from '../../../controls/ColumnInstruction.js'
+import { useLoadTable } from '../../../index.js'
+import type { StepComponentProps } from '../../../types.js'
 
 /**
  * Provides inputs for a step that needs lists of columns.
  */
-export const SpreadInputs: React.FC<StepComponentProps> = memo(
-	function SpreadInputs({ step, store, table, onChange, input }) {
-		const internal = useMemo(() => step as SpreadStep, [step])
+export const ColumnListInputs: React.FC<StepComponentProps> = memo(
+	function ColumnListInputs({ step, store, table, onChange, input }) {
+		const internal = useMemo(() => step as ColumnListStep, [step])
 
 		const tbl = useLoadTable(input || step.input, table, store)
-		const columns = useColumns(internal, onChange)
+
+		const columns = useColumns(internal, tbl, onChange)
 
 		const handleButtonClick = useCallback(() => {
 			onChange &&
@@ -33,29 +29,13 @@ export const SpreadInputs: React.FC<StepComponentProps> = memo(
 					...internal,
 					args: {
 						...internal.args,
-						to: [...internal.args.to, first(tbl)],
+						columns: [...internal.args.columns, first(tbl)],
 					},
 				})
 		}, [internal, tbl, onChange])
 
-		const handleColumnChange = useHandleDropdownChange(
-			step,
-			'args.column',
-			onChange,
-		)
-
 		return (
 			<Container>
-				<TableColumnDropdown
-					required
-					table={tbl}
-					label={'Column to spread'}
-					selectedKey={(step.args as SpreadArgs).column}
-					onChange={handleColumnChange}
-				/>
-
-				<Label>New column names</Label>
-
 				{columns}
 				<ActionButton
 					onClick={handleButtonClick}
@@ -73,31 +53,36 @@ function first(table?: ColumnTable): string {
 	return table?.columnNames()[0] as string
 }
 
-function useColumns(step: SpreadStep, onChange?: (step: Step) => void) {
+function useColumns(
+	step: ColumnListStep,
+	table?: ColumnTable,
+	onChange?: (step: Step) => void,
+) {
 	return useMemo(() => {
-		return (step.args.to || []).map((column: string, index: number) => {
+		return (step.args.columns || []).map((column: string, index: number) => {
 			const handleColumnChange = (col: string) => {
 				const update = { ...step }
-				set(update, `args.to[${index}]`, col)
+				set(update, `args.columns[${index}]`, col)
 				onChange && onChange(update)
 			}
 
 			const handleDeleteClick = () => {
 				const update = { ...step }
-				update.args.to.splice(index, 1)
+				update.args.columns.splice(index, 1)
 				onChange && onChange(update)
 			}
 
 			return (
-				<ColumnSpread
-					key={`column-list-${index}`}
+				<ColumnInstruction
+					key={`column-list-${column}-${index}`}
+					table={table}
 					column={column}
 					onChange={handleColumnChange}
 					onDelete={handleDeleteClick}
 				/>
 			)
 		})
-	}, [step, onChange])
+	}, [step, table, onChange])
 }
 
 const Container = styled.div`
