+++ javascript/react/src/verbs/shared/FilterFunction/FilterFunction.tsx
@@ -10,13 +10,13 @@ import {
 } from '@data-wrangling-components/core'
 import set from 'lodash-es/set.js'
 import { memo, useCallback, useMemo } from 'react'
-import { useLoadTable, useHandleDropdownChange } from '../../common/index.js'
-import type { StepComponentProps } from '../../types.js'
-import { ColumnOrValueComboBox } from '../ColumnOrValueComboBox'
+import { useLoadTable, useHandleDropdownChange } from '../../../common/index.js'
 import {
+	ColumnOrValueComboBox,
 	NumericComparisonOperatorDropdown,
 	StringComparisonOperatorDropdown,
-} from '../index.js'
+} from '../../../controls/index.js'
+import type { StepComponentProps } from '../../../types.js'
 
 /**
  * Just the comparison logic/ops for a filter.
