+++ javascript/react/src/verbs/shared/FilterFunction/index.ts
