+++ javascript/react/src/verbs/shared/index.ts
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './ColumnListInputs/index.js'
+export * from './FilterFunction/index.js'
+export * from './FilterInputs/index.js'
+export * from './JoinInputs/index.js'
